login-logout-room added

+++ studybud/base/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http  import HttpResponse
 from django.contrib  import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
-from django.http import HttpResponse
-from .models import Room,Topic
+from django.contrib.auth  import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room,Topic,Message
 from .forms import RoomForm
 
 # Create your views here.
@@ -14,16 +17,50 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
-    if request.method=="POST":
-        username=request.POST.get('username')
-        password=request.POST.get('password')
-    try:
-        user= User.objects.get(username=username)
-    except:
-        messages.error(request,'User does not exist')
-    context={}
-    return render(request,'base/login_registration.html',context)
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+        
+        user =authenticate(request,username=username, password=password)
+        
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'User OR Password does not exist')
+
+    context = {'page':page}
+    return render(request, 'base/login_register.html', context)
+  
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form =UserCreationForm()
     
+    if request.method == "POST":
+        form =UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occured during registration")
+            
+    return render(request, 'base/login_register.html',{'form':form})
 
 def home(request):
     q=request.GET.get('q') if request.GET.get('q') !=None else ''
@@ -41,9 +78,22 @@ def home(request):
 
 def room(request,pk):
     room=Room.objects.get(id=pk)
-    context={'room':room}
+    room_messages=room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    
+    if request.method == "POST":
+        messages = Message.objects.create(
+            user = request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room',pk=room.id)
+    context={'room':room,'room_messages':room_messages,'participants':participants}
     return render(request,'base/room.html',context)
 
+
+@login_required(login_url='login')    
 def createRoom(request):
     form =RoomForm()
     if request.method == 'POST':
@@ -55,10 +105,13 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html',context)
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
     
     if request.method =='POST':
         form=RoomForm(request.POST, instance=room)
@@ -70,11 +123,28 @@ def updateRoom(request,pk):
     
     return render(request, 'base/room_form.html',context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     
     return render(request, 'base/delete.html',{'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message = Message.objects.get(id=pk)
+    
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!')
+    
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    
+    return render(request, 'base/delete.html',{'obj':messages})
